Post-crud done

+++ src/main/java/com/codepresso/meu/service/PostService.java
@@ -12,8 +12,17 @@ import java.util.List;
 public class PostService {
     private PostMapper postMapper;
 
+    // test
+    public List<Post> getAllPost() {
+        return postMapper.findAll();
+    }
+
     public List<Post> getPostByFeed(Integer id) {
         return postMapper.findByFeed(id);
+    }
+
+    public Post getPostById(Integer id) {
+        return postMapper.findOne(id);
     }
 
     public Boolean savePost(Post post) {
@@ -21,12 +30,22 @@ public class PostService {
         return result == 1;
     }
 
-    public Boolean updatePost(Post post) {
+    public Boolean updatePost(Post post, Integer logInUserId) {
+        Post originalPost = postMapper.findOne(post.getPostId());
+        if(!originalPost.getUserId().equals(logInUserId)) {
+            return false;
+        }
+
         Integer result = postMapper.update(post);
         return result == 1;
     }
 
-    public Boolean deletePost(Integer id) {
+    public Boolean deletePost(Integer id, Integer logInUserId) {
+        Post originalPost = postMapper.findOne(id);
+        if(!originalPost.getUserId().equals(logInUserId)) {
+            return false;
+        }
+
         Integer result = postMapper.delete(id);
         return result == 1;
     }
